change the model definition of the Bet model. Odds is now type decimal

+++ server/db/models/bet.js
@@ -38,7 +38,7 @@ const Bet = db.define('bet', {
     allowNull: false
   },
   odds: {
-    type: Sequelize.INTEGER,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false
   },
   final: {
